requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.0
pydantic~=2.10.6
numpy~=2.2.4
openai~=1.68.2
anthropic~=0.49.0
sympy~=1.13.1
python-dotenv~=1.0.1
httpx>=0.26.0  # For async HTTP requests
pytest~=8.3.5
pytest-asyncio>=0.23.0  # For async tests 
torch~=2.6.0
transformers~=4.50.0
peft~=0.15.0
tokenizers>=0.13.3
safetensors>=0.3.1
accelerate>=0.20.3
dask>=2024.4.1
pandas>=2.0.3
scipy>=1.10.1
scikit-learn>=1.3.0
ray>=2.3.1
numba>=0.56.4 
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-anthropic>=0.0.1
sentence-transformers>=2.2.2
tqdm>=4.65.0
pillow>=9.5.0
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.14.1
ipywidgets>=8.0.6
streamlit>=1.22.0
typing_extensions>=4.5.0
mypy>=1.3.0
flake8>=6.0.0
black>=23.3.0
isort>=5.12.0
pre-commit>=3.3.2
pytest-cov>=4.1.0
loguru>=0.7.0
aiofiles>=23.1.0
async-timeout>=4.0.2 
